cascade-plot added & working
Also I initialized the title of the sn.plot() just to put something if title is not passed, because if not breaks and who cares for the title anyway in most of the cases.  The plot works! I will procide to try it for a higher repetitions number to see if it plots something better.